remove relaxed module, add anybitpattern
