--- src/relaxed.rs
@@ -1,203 +1,0 @@
-//! Versions of the casting functions exposed in crate root
-//! with slightly relaxed requirements, allowing use of the [`NoPadding`]
-//! trait.
-//!
-//! The functions exposed here are entirely equivalent to the ones exposed
-//! in crate main and *should* work in any case that those functions work
-//! since [`NoPadding`] is a superset of [`Pod`]. However, we haven't found
-//! a way to *prove* that there's no way to cause this to break existing code,
-//! so these versions are exposed here rather than replacing the originals, at least
-//! until the next major version of `bytemuck`.
-
-use crate::{internal, NoPadding, Pod, PodCastError};
-
-/// Re-interprets `&T` as `&[u8]`.
-///
-/// Any ZST becomes an empty slice, and in that case the pointer value of that
-/// empty slice might not match the pointer value of the input reference.
-#[inline]
-pub fn bytes_of<T: NoPadding>(t: &T) -> &[u8] {
-  unsafe { internal::bytes_of(t) }
-}
-
-/// Re-interprets `&mut T` as `&mut [u8]`.
-///
-/// Any ZST becomes an empty slice, and in that case the pointer value of that
-/// empty slice might not match the pointer value of the input reference.
-#[inline]
-pub fn bytes_of_mut<T: Pod>(t: &mut T) -> &mut [u8] {
-  unsafe { internal::bytes_of_mut(t) }
-}
-
-/// Re-interprets `&[u8]` as `&T`.
-///
-/// ## Panics
-///
-/// This is [`try_from_bytes`] but will panic on error.
-#[inline]
-pub fn from_bytes<T: NoPadding>(s: &[u8]) -> &T {
-  unsafe { internal::from_bytes(s) }
-}
-
-/// Re-interprets `&mut [u8]` as `&mut T`.
-///
-/// ## Panics
-///
-/// This is [`try_from_bytes_mut`] but will panic on error.
-#[inline]
-pub fn from_bytes_mut<T: Pod>(s: &mut [u8]) -> &mut T {
-  unsafe { internal::from_bytes_mut(s) }
-}
-
-/// Re-interprets `&[u8]` as `&T`.
-///
-/// ## Failure
-///
-/// * If the slice isn't aligned for the new type
-/// * If the slice's length isn’t exactly the size of the new type
-#[inline]
-pub fn try_from_bytes<T: NoPadding>(s: &[u8]) -> Result<&T, PodCastError> {
-  unsafe { internal::try_from_bytes(s) }
-}
-
-/// Re-interprets `&mut [u8]` as `&mut T`.
-///
-/// ## Failure
-///
-/// * If the slice isn't aligned for the new type
-/// * If the slice's length isn’t exactly the size of the new type
-#[inline]
-pub fn try_from_bytes_mut<T: Pod>(
-  s: &mut [u8],
-) -> Result<&mut T, PodCastError> {
-  unsafe { internal::try_from_bytes_mut(s) }
-}
-
-/// Cast `T` into `U`
-///
-/// ## Panics
-///
-/// * This is like [`try_cast`](try_cast), but will panic on a size mismatch.
-#[inline]
-pub fn cast<A: NoPadding, B: Pod>(a: A) -> B {
-  unsafe { internal::cast(a) }
-}
-
-/// Cast `&mut T` into `&mut U`.
-///
-/// ## Panics
-///
-/// This is [`try_cast_mut`] but will panic on error.
-#[inline]
-pub fn cast_mut<A: Pod, B: Pod>(a: &mut A) -> &mut B {
-  unsafe { internal::cast_mut(a) }
-}
-
-/// Cast `&T` into `&U`.
-///
-/// ## Panics
-///
-/// This is [`try_cast_ref`] but will panic on error.
-#[inline]
-pub fn cast_ref<A: NoPadding, B: Pod>(a: &A) -> &B {
-  unsafe { internal::cast_ref(a) }
-}
-
-/// Cast `&[A]` into `&[B]`.
-///
-/// ## Panics
-///
-/// This is [`try_cast_slice`] but will panic on error.
-#[inline]
-pub fn cast_slice<A: NoPadding, B: Pod>(a: &[A]) -> &[B] {
-  unsafe { internal::cast_slice(a) }
-}
-
-/// Cast `&mut [T]` into `&mut [U]`.
-///
-/// ## Panics
-///
-/// This is [`try_cast_slice_mut`] but will panic on error.
-#[inline]
-pub fn cast_slice_mut<A: Pod, B: Pod>(a: &mut [A]) -> &mut [B] {
-  unsafe { internal::cast_slice_mut(a) }
-}
-
-/// As `align_to`, but safe because of the [`Pod`] bound.
-#[inline]
-pub fn pod_align_to<T: NoPadding, U: Pod>(vals: &[T]) -> (&[T], &[U], &[T]) {
-  unsafe { vals.align_to::<U>() }
-}
-
-/// As `align_to_mut`, but safe because of the [`Pod`] bound.
-#[inline]
-pub fn pod_align_to_mut<T: Pod, U: Pod>(
-  vals: &mut [T],
-) -> (&mut [T], &mut [U], &mut [T]) {
-  unsafe { vals.align_to_mut::<U>() }
-}
-
-/// Try to cast `T` into `U`.
-///
-/// Note that for this particular type of cast, alignment isn't a factor. The
-/// input value is semantically copied into the function and then returned to a
-/// new memory location which will have whatever the required alignment of the
-/// output type is.
-///
-/// ## Failure
-///
-/// * If the types don't have the same size this fails.
-#[inline]
-pub fn try_cast<A: NoPadding, B: Pod>(a: A) -> Result<B, PodCastError> {
-  unsafe { internal::try_cast(a) }
-}
-
-/// Try to convert a `&T` into `&U`.
-///
-/// ## Failure
-///
-/// * If the reference isn't aligned in the new type
-/// * If the source type and target type aren't the same size.
-#[inline]
-pub fn try_cast_ref<A: Pod, B: Pod>(a: &A) -> Result<&B, PodCastError> {
-  unsafe { internal::try_cast_ref(a) }
-}
-
-/// Try to convert a `&mut T` into `&mut U`.
-///
-/// As [`try_cast_ref`], but `mut`.
-#[inline]
-pub fn try_cast_mut<A: Pod, B: Pod>(a: &mut A) -> Result<&mut B, PodCastError> {
-  unsafe { internal::try_cast_mut(a) }
-}
-
-/// Try to convert `&[A]` into `&[B]` (possibly with a change in length).
-///
-/// * `input.as_ptr() as usize == output.as_ptr() as usize`
-/// * `input.len() * size_of::<A>() == output.len() * size_of::<B>()`
-///
-/// ## Failure
-///
-/// * If the target type has a greater alignment requirement and the input slice
-///   isn't aligned.
-/// * If the target element type is a different size from the current element
-///   type, and the output slice wouldn't be a whole number of elements when
-///   accounting for the size change (eg: 3 `u16` values is 1.5 `u32` values, so
-///   that's a failure).
-/// * Similarly, you can't convert between a [ZST](https://doc.rust-lang.org/nomicon/exotic-sizes.html#zero-sized-types-zsts)
-///   and a non-ZST.
-#[inline]
-pub fn try_cast_slice<A: Pod, B: Pod>(a: &[A]) -> Result<&[B], PodCastError> {
-  unsafe { internal::try_cast_slice(a) }
-}
-
-/// Try to convert `&mut [A]` into `&mut [B]` (possibly with a change in
-/// length).
-///
-/// As [`try_cast_slice`], but `&mut`.
-#[inline]
-pub fn try_cast_slice_mut<A: Pod, B: Pod>(
-  a: &mut [A],
-) -> Result<&mut [B], PodCastError> {
-  unsafe { internal::try_cast_slice_mut(a) }
-}
